new linear approximation class
